+++ src/server/game/Entities/GameObject/GameObjectData.h
@@ -1115,7 +1115,7 @@ struct GameObjectData : public SpawnData
     QuaternionData rotation;
     uint32 animprogress = 0;
     GOState goState = GO_STATE_ACTIVE;
-    uint8 artKit = 0;
+    uint32 artKit = 0;
 };
 
 enum class GameObjectActions : uint32
